+++ src/frontend/App.tsx
@@ -3,14 +3,14 @@ import { Router } from '@reach/router';
 import styled, { ThemeProvider, createGlobalStyle } from 'styled-components';
 import moment from 'moment-timezone';
 
-import { NotFound } from './NotFound';
-import { FetchToken } from './FetchToken';
+import { NotFound } from './pages/NotFound';
+import { SSOLoginCallbackHandler } from './SSOLoginCallbackHandler';
 import { AuthContextWrapper } from './AuthContext';
 import { NotificationsContextWrapper } from './NotificationsContext';
 import { Navigation } from './Navigation';
 import { FrontPage } from './pages/FrontPage';
-import { Admin } from './Admin';
-import { Volunteer } from './Volunteer';
+import { AdminRouter } from './AdminRouter';
+import { VolunteerRouter } from './VolunteerRouter';
 
 import 'turretcss/turret/_color.css';
 import './assets/turret.css';
@@ -46,10 +46,10 @@ export const App = () => {
             <AppContainer className="container">
               <Router>
                 <NotFound default />
-                <Admin path="admin/*" />
-                <Volunteer path="volunteer/*" />
+                <AdminRouter path="admin/*" />
+                <VolunteerRouter path="volunteer/*" />
                 <FrontPage path="/" />
-                <FetchToken path="login/:nonce" />
+                <SSOLoginCallbackHandler path="login/:nonce" />
               </Router>
             </AppContainer>
           </NotificationsContextWrapper>
