+++ src/frontend/pages/Booking/BookingCalendar/shared-constants.ts
